:truck: Rename remote_powershell.py to remote_script.py
