+++ data_safe_haven/infrastructure/components/dynamic/__init__.py
@@ -2,7 +2,7 @@ from .azuread_application import AzureADApplication, AzureADApplicationProps
 from .blob_container_acl import BlobContainerAcl, BlobContainerAclProps
 from .compiled_dsc import CompiledDsc, CompiledDscProps
 from .file_share_file import FileShareFile, FileShareFileProps
-from .remote_powershell import RemoteScript, RemoteScriptProps
+from .remote_script import RemoteScript, RemoteScriptProps
 from .ssl_certificate import SSLCertificate, SSLCertificateProps
 
 __all__ = [
